+++ src/sql/cmd/snippets.py
@@ -2,6 +2,7 @@ from sql import util
 from sql.exceptions import UsageError
 from sql.cmd.cmd_utils import CmdParser
 from sql.store import store
+from sql import store_utils
 from sql.display import Table, Message
 
 
@@ -64,7 +65,7 @@ def snippets(others):
         help="Force delete all stored snippets",
         required=False,
     )
-    all_snippets = util.get_all_keys()
+    all_snippets = store_utils.get_all_keys()
     if len(others) == 1:
         if others[0] in all_snippets:
             return str(store[others[0]])
@@ -87,7 +88,7 @@ def snippets(others):
             return Table(["Stored snippets"], [[snippet] for snippet in all_snippets])
 
     if args.delete:
-        deps = util.get_key_dependents(args.delete)
+        deps = store_utils.get_key_dependents(args.delete)
         if deps:
             deps = ", ".join(deps)
             raise UsageError(
@@ -97,18 +98,18 @@ def snippets(others):
             )
         else:
             key = args.delete
-            remaining_keys = util.del_saved_key(key)
+            remaining_keys = store_utils.del_saved_key(key)
             return _modify_display_msg(key, remaining_keys)
 
     elif args.delete_force:
         key = args.delete_force
-        deps = util.get_key_dependents(key)
-        remaining_keys = util.del_saved_key(key)
+        deps = store_utils.get_key_dependents(key)
+        remaining_keys = store_utils.del_saved_key(key)
         return _modify_display_msg(key, remaining_keys, deps)
 
     elif args.delete_force_all:
-        deps = util.get_key_dependents(args.delete_force_all)
+        deps = store_utils.get_key_dependents(args.delete_force_all)
         deps.append(args.delete_force_all)
         for key in deps:
-            remaining_keys = util.del_saved_key(key)
+            remaining_keys = store_utils.del_saved_key(key)
         return _modify_display_msg(", ".join(deps), remaining_keys)
